AppsPlotting

Simple shiny app for viewing time series plots with box plots or line plots. You can use up to 3 categorical variables to filter by and one additional variable to filter as a bonus. You'll need to have shiny, shinyWidgets, htmltools

AppsPlotting(
  AzureCredsFile = NULL,
  HeaderColor = "blue",
  AppWidth = 12L,
  LogoWidth = "750px",
  LogoHeight = "100px",
  GroupVarsBoxColor = "navy",
  VarsBoxColor = "purple",
  FilterBoxColor = "blue",
  PlotBoxColor = "aqua",
  CreatePlotButtonColor = "royal",
  UpdatePlotButtonColor = "default",
  ResetPlotButtonColor = "default",
  H3Color = "darkblue",
  H4Color = "darkblue",
  AppTextColor = "blue",
  Browser = FALSE,
  Docker = FALSE,
  DockerPathToData = NULL,
  UserName_Password_DT = NULL,
  RunMode = "package",
  PlotObjectHome = NULL,
  Debug = FALSE
)

Arguments

AzureCredsFile
Path to your credentials file

HeaderColor
'black', 'blue', 'purple', 'green', 'red', 'yellow'

AppWidth
Width of boxes

GroupVarsBoxColor
Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'

VarsBoxColor
Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'

FilterBoxColor
Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'

PlotBoxColor
Choose from 'red', 'yellow', 'aqua', 'blue', 'light-blue', 'green', 'navy', 'teal', 'olive', 'lime', 'orange', 'fuchsia', 'purple', 'maroon', 'black'

CreatePlotButtonColor
Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'

UpdatePlotButtonColor
Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'

ResetPlotButtonColor
Choose from 'default', 'primary', 'warning', 'danger', 'success', 'royal'

H3Color
Header colors 'darkblue'

H4Color
Text below headers but not the input cells; 'blue2'

AppTextColor
'white'

Browser
FALSE

Docker
FALSE

DockerPathToData
bla

UserName_Password_DT
NULL. In order to enforce authentication, supply a data.table with columns 'UserName' which contains the names of your users and 'Password' which contains the acceptable passwords. E.g. data.table::data.table(UserName = c('Adrian Antico', 'Guest'), Password = c('Password1', 'Password2')). Case sensitivity applies.

RunMode
= 'package', 'local'

PlotObjectHome
NULL

Debug
FALSE

See also

Author

Adrian Antico

Examples

if (FALSE) {
# Pull Data
# data <- data.table::fread(system.file('tests/QA_DataSets/ThreeGroup-FC-Walmart-XREG3.csv', package = "RemixAutoML"))
# data[, Date := as.Date(Date)]

# Run App
# RemixAutoML::AppsPlotting(
#   HeaderColor = 'black',
#   AppWidth = 12L,
#   LogoWidth = '1000px',
#   LogoHeight = '100px',
#   GroupVarsBoxColor = 'navy',
#   VarsBoxColor = 'purple',
#   FilterBoxColor = 'blue',
#   PlotBoxColor = 'aqua',
#   CreatePlotButtonColor = 'default',
#   UpdatePlotButtonColor = 'default',
#   ResetPlotButtonColor = 'default',
#   H3Color = 'darkblue',
#   H5Color = 'blue2',
#   AppTextColor = 'white',
#   Docker = FALSE,
#   Browser = FALSE,
#   UserName_Password_DT = NULL,
#   PlotObjectHome = NULL,
#   Debug = FALSE)
#
# Debug = TRUE
# HeaderColor = 'black'
# AppWidth = 12L
# LogoWidth = '1000px'
# LogoHeight = '100px'
# GroupVarsBoxColor = 'navy'
# VarsBoxColor = 'purple'
# FilterBoxColor = 'blue'
# PlotBoxColor = 'aqua'
# CreatePlotButtonColor = 'default'
# UpdatePlotButtonColor = 'default'
# ResetPlotButtonColor = 'default'
# H3Color = 'darkblue'
# H4Color = 'blue2'
# AppTextColor = 'white'
# Docker = FALSE
# Browser = FALSE
# UserName_Password_DT = NULL
# Debug = TRUE
}