AutoBanditSarima

AutoBanditSarima is a multi-armed bandit model testing framework for SARIMA. Randomized probability matching is the underlying bandit algorithm. Model evaluation is done by blending the training error and the validation error from testing the model on out of sample data. The bandit algorithm compares the performance of the current build against the previous builds which starts with the classic auto.arima from the forecast package. Depending on how many lags, moving averages, seasonal lags and moving averages you test the number of combinations of features to test begins to approach 100,000 different combinations of settings. The function tests out transformations, differencing, and variations of the lags and moving averages. The paramter space is broken up into various buckets that are increasing in sophistication. The bandit algorithm samples from those buckets and based on many rounds of testing it determines which buckets to generate samples from more frequently based on the models performance coming from that bucket. All of the models have performance data collected on them and a final rebuild is initiated when a winner is found. The rebuild process begins by retraining the model with the settings that produced the best performance. If the model fails to build, for whatever reason, the next best buildable model is rebuilt.

AutoBanditSarima(
  data,
  FilePath = NULL,
  ByDataType = TRUE,
  TargetVariableName,
  DateColumnName,
  TimeAggLevel = "week",
  EvaluationMetric = "MAE",
  NumHoldOutPeriods = 5L,
  NumFCPeriods = 5L,
  MaxLags = 5L,
  MaxSeasonalLags = 0L,
  MaxMovingAverages = 5L,
  MaxSeasonalMovingAverages = 0L,
  MaxFourierPairs = 2L,
  TrainWeighting = 0.5,
  MaxConsecutiveFails = 25L,
  MaxNumberModels = 100L,
  MaxRunTimeMinutes = 10L,
  NumberCores = max(1L, min(4L, parallel::detectCores() - 2L)),
  DebugMode = FALSE
)

Arguments

data
Source data.table

FilePath
NULL to return nothing. Provide a file path to save the model and xregs if available

ByDataType
TRUE returns the best model from the four base sets of possible models. FALSE returns the best model.

TargetVariableName
Name of your time series target variable

DateColumnName
Name of your date column

TimeAggLevel
Choose from "year", "quarter", "month", "week", "day", "hour"

EvaluationMetric
Choose from MAE, MSE, and MAPE

NumHoldOutPeriods
Number of time periods to use in the out of sample testing

NumFCPeriods
Number of periods to forecast

MaxLags
A single value of the max number of lags to test

MaxSeasonalLags
A single value of the max number of seasonal lags to test

MaxMovingAverages
A single value of the max number of moving averages to test

MaxSeasonalMovingAverages
A single value of the max number of seasonal moving averages to test

MaxFourierPairs
A single value of the max number of fourier pairs to test

TrainWeighting
Model ranking is based on a weighted average of training metrics and out of sample metrics. Supply the weight of the training metrics, such as 0.50 for 50 percent.

MaxConsecutiveFails
When a new best model is found MaxConsecutiveFails resets to zero. Indicated the number of model attemps without a new winner before terminating the procedure.

MaxNumberModels
Indicate the maximum number of models to test.

MaxRunTimeMinutes
Indicate the maximum number of minutes to wait for a result.

NumberCores
Default max(1L, min(4L, parallel::detectCores()-2L))

DebugMode
Set to TRUE to get print outs of particular steps helpful in tracing errors

Value

data.table containing historical values and the forecast values along with the grid tuning results in full detail, as a second data.table

See also

Author

Adrian Antico

Examples