AutoH2oMLClassifier

AutoH2oMLClassifier is an automated H2O modeling framework with grid-tuning and model evaluation that runs a variety of steps. First, a stratified sampling (by the target variable) is done to create train and validation sets. Then, the function will run a random grid tune over N number of models and find which model is the best (a default model is always included in that set). Once the model is identified and built, several other outputs are generated: validation data with predictions, evaluation plot, evaluation metrics, variable importance, partial dependence calibration plots, and column names used in model fitting.

AutoH2oMLClassifier(
  OutputSelection = c("EvalMetrics", "PDFs", "Score_TrainData"),
  data = NULL,
  TrainOnFull = FALSE,
  ValidationData = NULL,
  TestData = NULL,
  TargetColumnName = NULL,
  FeatureColNames = NULL,
  ExcludeAlgos = NULL,
  eval_metric = "auc",
  CostMatrixWeights = c(1, 0, 0, 1),
  MaxMem = {     gc()    
    paste0(as.character(floor(as.numeric(system("awk '/MemFree/ {print $2}' /proc/meminfo",
    intern = TRUE))/1e+06)), "G") },
  NThreads = max(1, parallel::detectCores() - 2),
  MaxModelsInGrid = 2,
  model_path = NULL,
  metadata_path = NULL,
  ModelID = "FirstModel",
  NumOfParDepPlots = 3,
  ReturnModelObjects = TRUE,
  SaveModelObjects = FALSE,
  SaveInfoToPDF = TRUE,
  IfSaveModel = "mojo",
  H2OShutdown = TRUE,
  H2OStartUp = TRUE,
  DebugMode = FALSE
)

Arguments

OutputSelection
You can select what type of output you want returned. Choose from c("EvalMetrics", "PDFs", "Score_TrainData")

data
This is your data set for training and testing your model

TrainOnFull
Set to TRUE to train on full data

ValidationData
This is your holdout data set used in modeling either refine your hyperparameters.

TestData
This is your holdout data set. Catboost using both training and validation data in the training process so you should evaluate out of sample performance with this data set.

TargetColumnName
Either supply the target column name OR the column number where the target is located (but not mixed types). Note that the target column needs to be a 0 | 1 numeric variable.

FeatureColNames
Either supply the feature column names OR the column number where the target is located (but not mixed types)

ExcludeAlgos
"DRF","GLM","XGBoost","GBM","DeepLearning" and "Stacke-dEnsemble"

eval_metric
This is the metric used to identify best grid tuned model. Choose from "AUC" or "logloss"

CostMatrixWeights
A vector with 4 elements c(True Positive Cost, False Negative Cost, False Positive Cost, True Negative Cost). Default c(1,0,0,1),

MaxMem
Set the maximum amount of memory you'd like to dedicate to the model run. E.g. "32G"

NThreads
Set the number of threads you want to dedicate to the model building

MaxModelsInGrid
Number of models to test from grid options (1080 total possible options)

model_path
A character string of your path file to where you want your output saved

metadata_path
A character string of your path file to where you want your model evaluation output saved. If left NULL, all output will be saved to model_path.

ModelID
A character string to name your model and output

NumOfParDepPlots
Tell the function the number of partial dependence calibration plots you want to create.

ReturnModelObjects
Set to TRUE to output all modeling objects (E.g. plots and evaluation metrics)

SaveModelObjects
Set to TRUE to return all modeling objects to your environment

SaveInfoToPDF
Set to TRUE to print model insights to PDF

IfSaveModel
Set to "mojo" to save a mojo file, otherwise "standard" to save a regular H2O model object

H2OShutdown
Set to TRUE to shutdown H2O after running the function

H2OStartUp
Set to FALSE

DebugMode
Set to TRUE to print out steps taken

Value

Saves to file and returned in list: VariableImportance.csv, Model, ValidationData.csv, EvalutionPlot.png, EvaluationMetrics.csv, ParDepPlots.R a named list of features with partial dependence calibration plots, GridCollect, and GridList

See also

Author

Adrian Antico

Examples

# \donttest{
# Create some dummy correlated data with numeric and categorical features
data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.85,
  N = 1000L,
  ID = 2L,
  ZIP = 0L,
  AddDate = FALSE,
  Classification = TRUE,
  MultiClass = FALSE)

TestModel <- RemixAutoML::AutoH2oMLClassifier(
  OutputSelection = c("EvalMetrics", "PDFs", "Score_TrainData"),
  data,
  TrainOnFull = FALSE,
  ValidationData = NULL,
  TestData = NULL,
  TargetColumnName = "Adrian",
  FeatureColNames = names(data)[!names(data) %in% c("IDcol_1", "IDcol_2","Adrian")],
  ExcludeAlgos = NULL,
  eval_metric = "auc",
  CostMatrixWeights = c(1,0,0,1),
  MaxMem = {gc();paste0(as.character(floor(as.numeric(system("awk '/MemFree/ {print $2}' /proc/meminfo", intern=TRUE)) / 1000000)),"G")},
  NThreads = max(1, parallel::detectCores()-2),
  MaxModelsInGrid = 10,
  model_path = normalizePath("./"),
  metadata_path = normalizePath("./"),
  ModelID = "FirstModel",
  NumOfParDepPlots = 3,
  ReturnModelObjects = TRUE,
  SaveModelObjects = FALSE,
  SaveInfoToPDF = TRUE,
  IfSaveModel = "mojo",
  H2OShutdown = TRUE,
  H2OStartUp = TRUE,
  DebugMode = FALSE)

#>  Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         48 seconds 162 milliseconds 
#>     H2O cluster timezone:       America/Phoenix 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.32.1.4 
#>     H2O cluster version age:    3 months and 3 days  
#>     H2O cluster name:           H2O_started_from_R_Bizon_ktw237 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   248.00 GB 
#>     H2O cluster total cores:    64 
#>     H2O cluster allowed cores:  62 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, Algos, AutoML, Core V3, TargetEncoder, Core V4 
#>     R Version:                  R version 4.1.0 (2021-05-18) 
#> 
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> [1] "Cannot build a XGboost model - no backend found."
#>   |                                                                              |                                                                      |   0%Error in .h2o.doSafeREST(h2oRestApiVersion = h2oRestApiVersion, urlSuffix = urlSuffix,  : 
#>   Unexpected CURL error: Send failure: Connection was reset

#> Error in h2o.get_job(job_key, jobPollSuccess, jobIsRecoverable): Error in .h2o.doSafeREST(h2oRestApiVersion = h2oRestApiVersion, urlSuffix = urlSuffix,  : 
#>   Unexpected CURL error: Send failure: Connection was reset

# }