ModelDataPrep

This function replaces inf values with NA, converts characters to factors, and imputes with constants

ModelDataPrep(
  data,
  Impute = TRUE,
  CharToFactor = TRUE,
  FactorToChar = FALSE,
  IntToNumeric = TRUE,
  LogicalToBinary = FALSE,
  DateToChar = FALSE,
  IDateConversion = FALSE,
  RemoveDates = FALSE,
  MissFactor = "0",
  MissNum = -1,
  IgnoreCols = NULL
)

Arguments

data
This is your source data you'd like to modify

Impute
Defaults to TRUE which tells the function to impute the data

CharToFactor
Defaults to TRUE which tells the function to convert characters to factors

FactorToChar
Converts to character

IntToNumeric
Defaults to TRUE which tells the function to convert integers to numeric

LogicalToBinary
Converts logical values to binary numeric values

DateToChar
Converts date columns into character columns

IDateConversion
Convert IDateTime to POSIXct and IDate to Date types

RemoveDates
Defaults to FALSE. Set to TRUE to remove date columns from your data.table

MissFactor
Supply the value to impute missing factor levels

MissNum
Supply the value to impute missing numeric values

IgnoreCols
Supply column numbers for columns you want the function to ignore

Value

Returns the original data table with corrected values

See also

Author

Adrian Antico

Examples

if (FALSE) {
# Create fake data
data <- RemixAutoML::FakeDataGenerator(
  Correlation = 0.75,
  N = 250000L,
  ID = 2L,
  ZIP = 0L,
  FactorCount = 6L,
  AddDate = TRUE,
  Classification = FALSE,
  MultiClass = FALSE)

# Check column types
str(data)

# Convert some factors to character
data <- RemixAutoML::ModelDataPrep(
  data,
  Impute       = TRUE,
  CharToFactor = FALSE,
  FactorToChar = TRUE,
  IntToNumeric = TRUE,
  LogicalToBinary = FALSE,
  DateToChar   = FALSE,
  IDateConversion = FALSE,
  RemoveDates  = TRUE,
  MissFactor   = "0",
  MissNum      = -1,
  IgnoreCols   = c("Factor_1"))

# Check column types
str(data)
}