AutoLagRollStatsScoring

AutoLagRollStatsScoring Builds lags and a large variety of rolling statistics with options to generate them for hierarchical categorical interactions.

AutoLagRollStatsScoring(
  data,
  RowNumsID = "temp",
  RowNumsKeep = 1,
  Targets = NULL,
  HierarchyGroups = NULL,
  IndependentGroups = NULL,
  DateColumn = NULL,
  TimeUnit = "day",
  TimeUnitAgg = "day",
  TimeGroups = "day",
  TimeBetween = NULL,
  RollOnLag1 = 1,
  Type = "Lag",
  SimpleImpute = TRUE,
  Lags = NULL,
  MA_RollWindows = NULL,
  SD_RollWindows = NULL,
  Skew_RollWindows = NULL,
  Kurt_RollWindows = NULL,
  Quantile_RollWindows = NULL,
  Quantiles_Selected = NULL,
  Debug = FALSE
)

Arguments

data
A data.table you want to run the function on

RowNumsID
The name of your column used to id the records so you can specify which rows to keep

RowNumsKeep
The RowNumsID numbers that you want to keep

Targets
A character vector of the column names for the reference column in which you will build your lags and rolling stats

HierarchyGroups
A vector of categorical column names that you want to have generate all lags and rolling stats done for the individual columns and their full set of interactions.

IndependentGroups
Only supply if you do not want HierarchyGroups. A vector of categorical column names that you want to have run independently of each other. This will mean that no interaction will be done.

DateColumn
The column name of your date column used to sort events over time

TimeUnit
List the time aggregation level for the time between events features, such as "hour", "day", "weeks", "months", "quarter", or "year"

TimeUnitAgg
List the time aggregation of your data that you want to use as a base time unit for your features. E.g. "day",

TimeGroups
A vector of TimeUnits indicators to specify any time-aggregated GDL features you want to have returned. E.g. c("hour", "day","week","month","quarter","year"). STILL NEED TO ADD these '1min', '5min', '10min', '15min', '30min', '45min'

TimeBetween
Specify a desired name for features created for time between events. Set to NULL if you don't want time between events features created.

RollOnLag1
Set to FALSE to build rolling stats off of target columns directly or set to TRUE to build the rolling stats off of the lag-1 target

Type
List either "Lag" if you want features built on historical values or "Lead" if you want features built on future values

SimpleImpute
Set to TRUE for factor level imputation of "0" and numeric imputation of -1

Lags
A numeric vector of the specific lags you want to have generated. You must include 1 if WindowingLag = 1.

MA_RollWindows
A numeric vector of the specific rolling statistics window sizes you want to utilize in the calculations.

SD_RollWindows
A numeric vector of Standard Deviation rolling statistics window sizes you want to utilize in the calculations.

Skew_RollWindows
A numeric vector of Skewness rolling statistics window sizes you want to utilize in the calculations.

Kurt_RollWindows
A numeric vector of Kurtosis rolling statistics window sizes you want to utilize in the calculations.

Quantile_RollWindows
A numeric vector of Quantile rolling statistics window sizes you want to utilize in the calculations.

Quantiles_Selected
Select from the following c("q5", "q10", "q15", "q20", "q25", "q30", "q35", "q40", "q45", "q50", "q55", "q60"," q65", "q70", "q75", "q80", "q85", "q90", "q95")

Debug
Set to TRUE to get a print out of which step you are on

Value

data.table of original data plus created lags, rolling stats, and time between event lags and rolling stats

See also

Author

Adrian Antico

Examples

# \donttest{
# Create fake Panel Data----
Count <- 1L
for(Level in LETTERS) {
  datatemp <- RemixAutoML::FakeDataGenerator(
    Correlation = 0.75,
    N = 25000L,
    ID = 0L,
    ZIP = 0L,
    FactorCount = 0L,
    AddDate = TRUE,
    Classification = FALSE,
    MultiClass = FALSE)
  datatemp[, Factor1 := eval(Level)]
  if(Count == 1L) {
    data <- data.table::copy(datatemp)
  } else {
    data <- data.table::rbindlist(
      list(data, data.table::copy(datatemp)))
  }
  Count <- Count + 1L
}

# Create ID columns to know which records to score
data[, ID := .N:1L, by = "Factor1"]

#>             Adrian Independent_Variable1 Independent_Variable2
#>      1: 0.92028606           -0.05287617             0.9484975
#>      2: 0.59456676           -1.33106358             0.2641961
#>      3: 0.81580387           -0.35187512             0.7033680
#>      4: 0.49911982           -1.49838142             0.2234916
#>      5: 0.48969516           -0.37787247             0.6853179
#>     ---                                                       
#> 649996: 0.71277890           -0.43999505             0.6440396
#> 649997: 0.06247416           -0.49714600             0.6082642
#> 649998: 0.26413942           -1.61794840             0.1983051
#> 649999: 0.82401290           -1.12104584             0.3259387
#> 650000: 0.61609174           -0.79273006             0.4526075
#>         Independent_Variable3 Independent_Variable4 Independent_Variable5
#>      1:              2.581827             13.221277             0.9739083
#>      2:              1.302384              3.678053             0.5140001
#>      3:              2.020546              7.542445             0.8386703
#>      4:              1.250435              3.491862             0.4727490
#>      5:              1.984403              7.274700             0.8278393
#>     ---                                                                  
#> 649996:              1.904157              6.713748             0.8025208
#> 649997:              1.837239              6.279181             0.7799129
#> 649998:              1.219334              3.384934             0.4453146
#> 649999:              1.385330              3.996146             0.5709104
#> 650000:              1.572407              4.818231             0.6727611
#>         Independent_Variable6 Independent_Variable7 Independent_Variable8
#>      1:             0.9947263             0.9868679             0.9611189
#>      2:             0.8753720             0.7169380             0.3685062
#>      3:             0.9654244             0.9157895             0.7680455
#>      4:             0.8608473             0.6875674             0.3250468
#>      5:             0.9629178             0.9098567             0.7532152
#>     ---                                                                  
#> 649996:             0.9569544             0.8958352             0.7189264
#> 649997:             0.9515009             0.8831268             0.6887620
#> 649998:             0.8506157             0.6673190             0.2971669
#> 649999:             0.8939508             0.7555862             0.4313720
#> 650000:             0.9237877             0.8202201             0.5518122
#>         Independent_Variable9 Independent_Variable10   DateTime Factor1    ID
#>      1:            0.89964743            0.809365497 1953-05-01       A 25000
#>      2:            0.06979959            0.004871983 1953-05-02       A 24999
#>      3:            0.49472648            0.244754290 1953-05-03       A 24998
#>      4:            0.04994850            0.002494852 1953-05-04       A 24997
#>      5:            0.46966061            0.220581091 1953-05-05       A 24996
#>     ---                                                                      
#> 649996:            0.41478702            0.172048268 2021-10-06       Z     5
#> 649997:            0.36998530            0.136889121 2021-10-07       Z     4
#> 649998:            0.03932492            0.001546450 2021-10-08       Z     3
#> 649999:            0.10623606            0.011286100 2021-10-09       Z     2
#> 650000:            0.20485352            0.041964963 2021-10-10       Z     1

data.table::set(data, i = which(data[["ID"]] == 2L), j = "ID", value = 1L)

# Score records
data <- RemixAutoML::AutoLagRollStatsScoring(

  # Data
  data                 = data,
  RowNumsID            = "ID",
  RowNumsKeep          = 1,
  DateColumn           = "DateTime",
  Targets              = "Adrian",
  HierarchyGroups      = c("Store","Dept"),
  IndependentGroups    = NULL,

  # Services
  TimeBetween          = NULL,
  TimeGroups           = c("days", "weeks", "months"),
  TimeUnit             = "day",
  TimeUnitAgg          = "day",
  RollOnLag1           = TRUE,
  Type                 = "Lag",
  SimpleImpute         = TRUE,

  # Calculated Columns
  Lags                  = list("days" = c(seq(1,5,1)),
                               "weeks" = c(seq(1,3,1)),
                               "months" = c(seq(1,2,1))),
  MA_RollWindows        = list("days" = c(seq(1,5,1)),
                               "weeks" = c(seq(1,3,1)),
                               "months" = c(seq(1,2,1))),
  SD_RollWindows        = list("days" = c(seq(1,5,1)),
                               "weeks" = c(seq(1,3,1)),
                               "months" = c(seq(1,2,1))),
  Skew_RollWindows      = list("days" = c(seq(1,5,1)),
                               "weeks" = c(seq(1,3,1)),
                               "months" = c(seq(1,2,1))),
  Kurt_RollWindows      = list("days" = c(seq(1,5,1)),
                               "weeks" = c(seq(1,3,1)),
                               "months" = c(seq(1,2,1))),
  Quantile_RollWindows  = list("days" = c(seq(1,5,1)),
                               "weeks" = c(seq(1,3,1)),
                               "months" = c(seq(1,2,1))),
  Quantiles_Selected    = c("q5","q10","q95"),
  Debug                 = FALSE)

#> Error: Some items of .SDcols are not column names: [Store, Dept]

# }